Decouple implementation from GUI

+++ Code/AMCore/AMGUI/windows/AMFramework/AMFramework/Controller/Controller_DBS_Projects.cs
@@ -241,7 +241,7 @@ namespace AMFramework.Controller
             {
                 _selectedProject = value;
                 Controller_Global.MainControl?.Show_Notification("Project", "New Project has been selected", 
-                                                                FontAwesome.WPF.FontAwesomeIcon.Info,
+                                                                (int)FontAwesome.WPF.FontAwesomeIcon.Info,
                                                                 null, null,null);
                 SelectProject(_selectedProject.ID);
             }
